Handles failures & refactors task image
